migration/requirements.txt
psycopg2-binary==2.9.9
supabase==2.17.0
python-dotenv==1.1.1
pandas==2.3.1
tqdm==4.67.1